feat: extend bitnet-rs for production-ready general LLMs
- Added `train.py` and `ds_config.json` for fine-tuning using Hugging Face `Trainer` and DeepSpeed.
- Updated `HuggingFaceLoader` in `bitnet-models` to natively load FP16 and FP32 SafeTensors directly, mapping them correctly for inference.
- Implemented `train-tokenizer` CLI subcommand in `bitnet-cli` to build and train BPE vocabularies using the `tokenizers` crate.
- Added `FP16` and `INT8` to `QuantizationType` and added a `--precision` flag across `bitnet-cli` commands.
- Configured `/v1/generate` alias in `bitnet-server` pointing to the enhanced inference handler to provide a standard inference REST endpoint.
- Updated documentation and addressed pipeline Clippy warnings.

+++ crates/bitnet-logits/tests/logits_proptests.rs
@@ -84,7 +84,7 @@ proptest! {
     ) {
         let mut logits = vec![val; n];
         softmax_in_place(&mut logits);
-        let expected = 1.0 / n as f32;
+        let expected = 1.0 / n as f64 as f32;
         for &p in &logits {
             prop_assert!(
                 (p - expected).abs() < 1e-4,
